+++ CHANGELOG.md
@@ -35,6 +35,7 @@
    - `<LexiconExtension>`, `<Extends>`, `<ExternalSynset>`,
      `<ExternalLexicalEntry>`, `<ExternalSense>`
    - `subcat` on `<Sense>`
+   - `members` on `<Synset>` ([#7])
 
 ### Fixed
 
@@ -85,6 +86,7 @@
 * Added a table for lexicon dependencies ([#7], [#89])
 * Added a table for lexicon extensions ([#99])
 * Added `logo` column to `lexicons` table ([#89])
+* Added a `synset_rank` column to `senses` table ([#89])
 
 
 ## [v0.5.1]